Не могу создать массив фото
