+++ js/data.js
@@ -1,7 +1,7 @@
 'use strict'
 
 // Обьявляем константы для задания
-const INITIAL_POST_COUNT = 1; // по-умаолчанию надо 25, для теста поставил 2
+const INITIAL_POST_COUNT = 25; // по-умаолчанию надо 25, для теста поставил 2
 
 const DESCRIPTION_LIST = [
   'Жить только для себя — есть злоупотребление.',
