+++ js/mini-photo.js
@@ -1,3 +1,28 @@
-const test = document.querySelector('.picture');
+'use strict'
+import {createPost} from './create-post.js';
+import {} from './data.js'
 
-console.log(typeof test)
+//Место, куда будем накидываьт мини фото
+const miniPhotoList = document.querySelector('.pictures__title')
+
+
+//Находим заготовленный шабло и достаем из него вложенную структуру
+const templateMiniPhoto = document.querySelector('#picture')
+  .content
+  .querySelector('.picture');
+
+//Создаем масив мини-фото, который будет вставлен на страницу
+const miniPhotos = createPost();
+
+
+//создаем
+miniPhotos.forEach((photo) => {
+  const miniPhotoElement = templateMiniPhoto.cloneNode(true);
+  miniPhotoElement.querySelector('.picture__img').textContent = po
+  miniPhotoList.appendChild(miniPhotoElement);
+
+});
+
+
+
+
